Added ico file and shortcut script (changes to readme)
The run_gmp.bat file could be used as a shortcut but you cannot change the icon of the shortcut if you transfer the file directly. You would instead have to create a shortcut from the run_gmp.bat file and use that instead on the desktop (or wherever you'd like to place the shortcut). From there you are given more freedom to edit the shortcut properties. Then you can use the ico file within this repository as the icon for the shortcut.